requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.1.4
pillow==10.3.0
scikit-learn==1.5.0
ipykernel==6.29.0
